Fix KPrint/KRun with default temporary directory (https://github.com/runtimeverification/pyk/pull/50)

+++ pyk/src/integration_tests/krun_test.py
@@ -2,7 +2,7 @@ import os
 import shutil
 from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import Iterable
+from typing import Iterable, Optional
 
 from pyk.ktool import KRun
 from pyk.ktool.kprint import SymbolTable
@@ -11,19 +11,23 @@ from .kompiled_test import KompiledTest
 
 
 class KRunTest(KompiledTest, ABC):
-    KRUN_USE_DIR: str = '.krun'
+    KRUN_USE_DIR: Optional[str] = '.krun'
     KRUN_INCLUDE_DIRS: Iterable[str] = []
 
     def setUp(self) -> None:
         super().setUp()
 
-        shutil.rmtree(self.KRUN_USE_DIR, ignore_errors=True)
-        os.makedirs(self.KRUN_USE_DIR)
+        if self.KRUN_USE_DIR:
+            shutil.rmtree(self.KRUN_USE_DIR, ignore_errors=True)
+            os.makedirs(self.KRUN_USE_DIR)
 
-        self.krun = KRun(self.kompiled_dir, use_directory=Path(self.KRUN_USE_DIR))
+            self.krun = KRun(self.kompiled_dir, use_directory=Path(self.KRUN_USE_DIR))
+        else:
+            self.krun = KRun(self.kompiled_dir)
 
     def tearDown(self) -> None:
-        shutil.rmtree(self.KRUN_USE_DIR, ignore_errors=True)
+        if self.KRUN_USE_DIR:
+            shutil.rmtree(self.KRUN_USE_DIR, ignore_errors=True)
         super().tearDown()
 
     @staticmethod
